Support Windows Build
On Windows, it needs `clang-cl` to build. Also, to support
`clang-cl` on bazel, we have to upgrade bazel version to `0.20.0`.
Then old version of `grpc` is not correspondent to new bazel.
So we upgraded the version, which leads to upgrade the version of
`protobuf` higher or equal to `3.6.0`.
Besides that, `boring-ssl` which google maintains for ssl, has a
problem on windows, which related to division __uint128_t. If you want
to look into it, please refer to the
[link](https://bugs.chromium.org/p/chromium/issues/detail?id=787617).