Redesign channel
Previously Channel class was designed mainly for sending and
receiving one type of message. So it was made as a template class.
But on demand of sending and receiving different kinds of types
for one channel, it now is made as a class sending and receiving
std::string. To send or recieve, now MessageSender<T> and
MessageReceiver<T> are introduced.

Fix bug: Heartbeat listener timeout callback after KillSelf

Optimize: when publishing, subscribing, sending and receiving through
the class mentioned above, MessageSender<T> and MessageReceiver<T>,
no copy happened. It writes to or reads from ChannelBuffer.

Enable c++ service using ROS service protocol